+++ lib/guardian.ex
@@ -340,16 +340,20 @@ defmodule Guardian do
       the_otp_app = unquote(otp_app)
       the_opts = unquote(opts)
 
-      # Provide a way to get at the configuration during compile time
-      # for other macros that may want to use them
-      @config fn ->
-        the_otp_app |> Application.compile_env(__MODULE__, []) |> Keyword.merge(the_opts)
-      end
-      @config_with_key fn key ->
-        @config.() |> Keyword.get(key) |> Guardian.Config.resolve_value()
-      end
-      @config_with_key_and_default fn key, default ->
-        @config.() |> Keyword.get(key, default) |> Guardian.Config.resolve_value()
+      # Provide a way to get at the permissions during compile time. Uses
+      # permissions from config if they are available and falls back to the
+      # permissins defined on the `use Guardian` implementation
+      #
+      # NOTE: Generally you can't use compile_env for most keys because that
+      # would prevent people from changing them at runtime for differen
+      # environements.And hardcoding secret keys wouldn't be considered a good
+      # practice.
+      @config_permissions fn ->
+        perms =
+          Application.compile_env(the_otp_app, [__MODULE__, :permissions]) ||
+            Keyword.get(the_opts, :permissions, [])
+
+        Guardian.Config.resolve_value(perms)
       end
 
       @doc """